move custom fleet handling to fleet_core
